made many changes condensed things a lot
NB: note to all please change the filepaths in each to where it is saved on your computer
