+++ .gitignore
@@ -1,3 +1,4 @@
 .idea
 venv
 
+/data/~$solvay_attributes.xlsx
