+++ solvayProcess/topDown/qReaction.py
@@ -1,6 +1,9 @@
-from machines import *
 from SolvayIO import *
 
+if __name__ == '__main__':
+    my_calciner = calciner.calciner_att()
+    emissions = calciner.qMachine(my_calciner)
+    print(emissions)
