+++ solvayProcess/topDown/reactantFlow.py
@@ -1,16 +1,9 @@
-from SolvayIO.reactantsIO.ammonia import ammonia_att
-from SolvayIO.reactantsIO.brine import brine_att
-from SolvayIO.reactantsIO.coke import coke_att
-from SolvayIO.reactantsIO.limestone import limestone_att
-from solvayProcess.machines import *
+from SolvayIO import *
 
 
-if __name__ == '__main__':
-    myBrine = brine_att
-    myAmmonia = ammonia_att
-    myCoke = coke_att
-    myLimestone = limestone_att
-    saturated_solution =
+
+
+
